blueprint3d/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow==10.4.0
httpx==0.27.2
